Niti_Network/backend/requirements.txt
# ─── Web Framework ─────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─── Database ──────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# ─── Auth / Security ───────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0

# ─── Data Validation ───────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ─── Object Storage ────────────────────────────────────────
minio==7.2.7

# ─── Vector DB (Biometrics) ────────────────────────────────
pymilvus==2.3.7

# ─── Biometrics ────────────────────────────────────────────
deepface==0.0.93
opencv-python-headless==4.9.0.80
numpy==1.26.4

# ─── Fraud / ML ────────────────────────────────────────────
scikit-learn==1.4.2
pandas==2.2.2
pyyaml==6.0.1

# ─── Monitoring ────────────────────────────────────────────
prometheus-fastapi-instrumentator==6.1.0

# ─── Utilities ─────────────────────────────────────────────
python-dotenv==1.0.1
aiofiles==23.2.1
structlog==24.1.0
